+++ sdk/src/main/java/com/mapbox/maps/renderer/widget/Widget.kt
@@ -10,12 +10,34 @@ abstract class Widget internal constructor() {
   internal abstract val renderer: WidgetRenderer
 
   /**
-   * Set absolute translation of widget in pixels.
+   * Update the widget to the new position.
    */
+  abstract fun setPosition(widgetPosition: WidgetPosition)
+
+  /**
+   * Get the current position of the widget.
+   */
+  abstract fun getPosition(): WidgetPosition
+
+  /**
+   * Set the translation of the widget in pixels, relative to it's current position.
+   *
+   * @param translateX the offset in pixels towards the right of the screen.
+   * @param translateY the offset in pixels towards the bottom of the screen.
+   */
+  @Deprecated(
+    message = "setTranslation is deprecated, please use setPosition instead.",
+    replaceWith = ReplaceWith("setPosition")
+  )
   abstract fun setTranslation(translateX: Float, translateY: Float)
 
   /**
-   * Set absolute rotation of widget in angles.
+   * Set the absolute rotation of widget in degrees.
    */
   abstract fun setRotation(angleDegrees: Float)
+
+  /**
+   * Get absolute rotation of widget in degrees.
+   */
+  abstract fun getRotation(): Float
 }