+++ sdk/src/test/java/com/mapbox/maps/EqualsHashCodeTest.kt
@@ -1,6 +1,7 @@
 package com.mapbox.maps
 
 import com.mapbox.maps.renderer.RendererError
+import com.mapbox.maps.renderer.widget.WidgetPosition
 import com.mapbox.maps.viewannotation.ViewAnnotationUpdateMode
 import nl.jqno.equalsverifier.EqualsVerifier
 import org.junit.Test
@@ -18,6 +19,16 @@ class EqualsHashCodeTest {
   fun `ViewAnnotationUpdateMode hashCode and equals test`() {
     val clazz = ViewAnnotationUpdateMode::class.java
     EqualsVerifier.forClass(clazz)
+      .usingGetClass()
+      .verify()
+  }
+
+  @Test
+  @OptIn(MapboxExperimental::class)
+  fun `WidgetPosition hashCode and equals test`() {
+    val clazz = WidgetPosition::class.java
+    EqualsVerifier.forClass(clazz)
+      .withIgnoredFields("horizontal", "vertical")
       .usingGetClass()
       .verify()
   }